firebase and clerk integrated, 1st attempt. Used their easier method with direct integration on dashboard, initialized auth in firebase

+++ config/firebaseConfig.js
@@ -10,8 +10,7 @@ import {
     doc,
     deleteDoc,
 } from "firebase/firestore";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
+import { getAuth } from "firebase/auth";
 
 // Your web app's Firebase configuration
 const firebaseConfig = {
@@ -27,5 +26,6 @@ const firebaseConfig = {
 
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
+const auth = getAuth(app);
 
-export { db, collection, addDoc, query, where, getDocs, doc, deleteDoc };
+export { db, collection, addDoc, query, where, getDocs, doc, deleteDoc, auth };
